Speed-up of AbelianGrouper.group_subops (Qiskit/qiskit-aqua#923)
* fast operator grouping with numpy

* add a trick for networkx

* remove unnecessary "[]"

* fix comments

* (wip) add test_abelian_grouper.py

* implement a fast pass of grouping

* fix lint

* implement a numpy-based coloring

* move abelian groupter tests

* simplify

* move abelian groupter tests to a separete file

* replace assert with assertTrue

* Fix lint

+++ qiskit/aqua/operators/primitive_ops/pauli_op.py
@@ -248,8 +248,8 @@ class PauliOp(PrimitiveOp):
         if not isinstance(other_op, PauliOp):
             return False
         # Don't use compose because parameters will break this
-        self_bits = self.primitive.z.astype(int) + 2 * self.primitive.x.astype(int)
-        other_bits = other_op.primitive.z.astype(int) + 2 * other_op.primitive.x.astype(int)
+        self_bits = self.primitive.z + 2 * self.primitive.x
+        other_bits = other_op.primitive.z + 2 * other_op.primitive.x
         return all((self_bits * other_bits) * (self_bits - other_bits) == 0)
 
     def to_circuit(self) -> QuantumCircuit:
